fix(resampler): explicit conversions to silence MSVC /W4 warnings
No behaviour change. memset() takes an int fill value, so 0.0f becomes 0;
std::floor() results and a size_t index are cast to int explicitly, as the
SIMD path in LanczosResampler.h already did.

+++ dsp/ResamplingContainer/Dependencies/LanczosResampler.h
@@ -308,7 +308,7 @@ private:
   inline void ReadSamples(double xBack, T** outputs, int s) const
   {
     double bufferReadPosition = mWritePos - xBack;
-    int bufferReadIndex = std::floor(bufferReadPosition);
+    int bufferReadIndex = static_cast<int>(std::floor(bufferReadPosition));
     double bufferFracPosition = 1.0 - (bufferReadPosition - bufferReadIndex);
 
     bufferReadIndex = (bufferReadIndex + kBufferSize) & (kBufferSize - 1);
